added simple @OneToMany relation

+++ src/main/java/com/de/brightslearning/football/demo/ClubRepository.java
@@ -2,5 +2,5 @@ package com.de.brightslearning.football.demo;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface ClubRepository extends CrudRepository<Club, Integer> {
+public interface ClubRepository extends CrudRepository<Club, Long> {
 }
